Add CpuLevel enum
